api/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.1
geopandas==0.14.1
fiona==1.9.5
pyproj==3.6.1
python-dotenv==1.0.0
pydantic-settings